+++ deepdoc_client_action/CHANGELOG.md
@@ -21,3 +21,4 @@
 ## 0.0.7
 - Added document URL support
 - Added pagination for document list by job
+- Modified list_documents to respond t an 'all' flag which returns all docs